I refactored the main method to make it work with the model, and fixed the algorithms (now they calculate a double and cast to an int to display the time).

+++ src/paystation/domain/ProgressiveRateStrategy.java
@@ -18,13 +18,13 @@ public class ProgressiveRateStrategy implements Strategy{
         }
         
         else if ((150 <= amount) && (amount < 350)) {
-            time = (amount-150)*(3/10) + 60;
+            time = (int)((amount-150)*(3.0/10.0) + 60);
         }
         
         else {
-            time = (amount-350)/5 + 120;
+            time = (int)((amount-350)/5.0 + 120);
         }
-    
+        
         return time;
     }
 }
